pyproject.toml
[project]
name = "bbdd_vec_ct"
version = "0.1.0"
description = " "
readme = "README.md"
requires-python = "==3.12.11"

dependencies = [
    "fastapi>=0.117.1",
    "google-genai>=1.38.0",
    "grpcio==1.60.1",
    "langchain>=0.3.30",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.30",
    "langchain-google-genai>=2.1.12",
    "langchain-huggingface>=0.3.1",
    "langchain-qdrant>=0.2.1",
    "pypdf>=6.1.0",
    "uvicorn>=0.36.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    "qdrant-client>=1.7.0",
    "tenacity>=8.0.0",
]